requirements.txt
requests>=2.31,<3
pydantic>=2.5,<3
python-dotenv>=1.0,<2
fastapi>=0.109,<1
uvicorn[standard]>=0.27,<1
